style fixes and moved layout and seo into component dirs
