+++ learning/web/src/templates/page.js
@@ -2,8 +2,8 @@ import React from "react"
 import { graphql } from "gatsby"
 
 import { MDXRenderer } from "gatsby-plugin-mdx"
-import Layout from "../components/layout"
-import SEO from "../components/seo"
+import Layout from "../components/Layout"
+import SEO from "../components/SEO"
 import { components } from "../utils/mdx-components"
 import { rhythm } from "../utils/typography"
 import { MDXProvider } from "@mdx-js/react"
